Question about reorder_vertices.jl
Hello :) 

Thank you very much for the code. I have a question about the vertex ordering for the CommonVertex case. 
The orders Pt = [P, A1, A2] and Ps = [P, B1, B2] is proposed in the book of Sauter and Schwab. Since this package uses the triangle representation x = x[3] + u*(x[1]-x[3]) + v*(x[2]-x[3]), is it possible that Pt = [A1, A2, P] and Ps = [B1, B2, P] might be the choice for the orders? 
In this case, the parameters u=0 and v=0 would point to the same node in both triangles. 
As I said, I am not completely sure, but I copied my proposal into the code. 

Best regards!